Refactor: 기존 hooks > useClientSearchParams 로 쓰던것을 refactora 및 분리
